# backend/backend_app/static/openapi.yaml
openapi: "3.0.3"
info:
  title: API do Petshop
  description: API responsável pelo gerenciamento de usuários, pets e serviços
  version: 1.0.0
  contact: 
    name: Tiago F. Hinterholz
    url: https://github.com/tiagohinterholz/Petshop_API
    email: fh.tiago@gmail.com
servers:
  - url: http://localhost:5000
...
paths:
...
  /clients:
    post:
      tags:
        - 'Clients'
      operationId: cadastrarClient
      summary: Cadastro de Clientes
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/ClientsRequest'
      responses:
        '201':
         description: "Client cadastrado com sucesso"
         content:
          'application/json':
            schema:
              $ref: '#/components/schemas/ClientsResponse'
        '400':
          description: "Problema nos dados de cadastro"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuário não autenticado para esta operação"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
        '500':
          description: "Erro inesperado ao cadastrar client"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - 'Clients'
      operationId: listarClients
      summary: Listagem de Clientes
      security:
        - JWT: []
      responses:
        '200':
          description: "Listagem de clientes realizada com sucesso"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientsResponse'
        '401':
          description: 'Usuário não autorizado'
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'    
  /clients/{client_id}:
    get:
      tags:
        - 'Clients'
      operationId: detalharClient
      summary: "Detalhar Cliente pelo client_id"
      parameters:
        - name: client_id
          in: path
          description: "ID do cliente a ser detalhado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Cliente detalhado com sucesso"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientsResponse"
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Usuário não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'   
    put:
      tags:
        - 'Clients'
      summary: "Atualizar cadastro do Cliente pelo client_id"
      operationId: atualizarClient
      parameters:
        - name: client_id
          in: path
          description: "ID do cliente a ser atualizado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/ClientsRequest"
      responses:
        '200':
          description: "Client atualizado com sucesso"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ClientsResponse'
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Usuario não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
    delete:
      tags:
        - 'Clients'
      operationId: deletarClient
      summary: "Deletar Cliente pelo client_id"
      parameters:
        - name: client_id
          in: path
          description: "ID do cliente a ser deletado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Client deletado com sucesso"
        '401':
          description: "User não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Client não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro ao excluir cliente"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
...
  /pets:
    post:
      tags:
        - 'Pets'
      operationId: cadastrarPet
      summary: Cadastro de Pets
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/PetsRequest'
      responses:
        '201':
         description: "Pet cadastrado com sucesso"
         content:
          'application/json':
            schema:
              $ref: '#/components/schemas/PetsResponse'
        '400':
          description: "Problema nos dados de cadastro"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuário não autenticado para esta operação"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
        '500':
          description: "Erro inesperado ao cadastrar pet"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - 'Pets'
      operationId: listarPets
      summary: Listagem de Pets
      security:
        - JWT: []
      responses:
        '200':
          description: "Listagem de pets realizada com sucesso"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PetsResponse'
        '401':
          description: 'Usuário não autorizado'
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pets/{pet_id}:
    get:
      tags:
        - 'Pets'
      operationId: detalharPet
      summary: "Detalhar Pet pelo pet_id"
      parameters:
        - name: pet_id
          in: path
          description: "ID do pet a ser detalhado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Pet detalhado com sucesso"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetsResponse"
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Pet não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'   
    put:
      tags:
        - 'Pets'
      summary: "Atualizar cadastro de Pet pelo pet_id"
      operationId: atualizarPet
      parameters:
        - name: pet_id
          in: path
          description: "ID do pet a ser atualizado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/PetsRequest"
      responses:
        '200':
          description: "Pet atualizado com sucesso"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PetsResponse'
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Pet não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
    delete:
      tags:
        - 'Pets'
      operationId: deletarPet
      summary: "Deletar Pet pelo pet_id"
      parameters:
        - name: pet_id
          in: path
          description: "ID do pet a ser deletado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Pet deletado com sucesso"
        '401':
          description: "User não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Pet não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro ao excluir Pet"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
  /contacts:
    post:
      tags:
        - 'Contacts'
      operationId: cadastrarContact
      summary: Cadastro de Contatos
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/ContactsRequest'
      responses:
        '201':
         description: "Contact cadastrado com sucesso"
         content:
          'application/json':
            schema:
              $ref: '#/components/schemas/ContactsResponse'
        '400':
          description: "Problema nos dados de cadastro"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuário não autenticado para esta operação"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
        '500':
          description: "Erro inesperado ao cadastrar contact"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - 'Contacts'
      operationId: listarContacts
      summary: Listagem de Contatos
      security:
        - JWT: []
      responses:
        '200':
          description: "Listagem de contatos realizada com sucesso"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactsResponse'
        '401':
          description: 'Usuário não autorizado'
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contacts/{contact_id}:
    get:
      tags:
        - 'Contacts'
      operationId: detalharContact
      summary: "Detalhar Contato pelo contact_id"
      parameters:
        - name: contact_id
          in: path
          description: "ID do contato a ser detalhado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Contato detalhado com sucesso"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactsResponse"
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Contato não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'   
    put:
      tags:
        - 'Contacts'
      summary: "Atualizar cadastro de Contato pelo contact_id"
      operationId: atualizarContact
      parameters:
        - name: contact_id
          in: path
          description: "ID do contato a ser atualizado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/ContactsRequest"
      responses:
        '200':
          description: "Contact atualizado com sucesso"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ContactsResponse'
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Contato não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
    delete:
      tags:
        - 'Contacts'
      operationId: deletarContact
      summary: "Deletar Contato pelo contact_id"
      parameters:
        - name: contact_id
          in: path
          description: "ID do contato a ser deletado"
          required: true
          schema:
            type: integer
      security:
        - JWT: []
      responses:
        '200':
          description: "Contato deletado com sucesso"
        '401':
          description: "User não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Client não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro ao excluir cliente"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    ClientsRequest:
      type: object
      required:
        - cpf
      properties:
        cpf:
          type: string
          maxLength: 14
          example: "000.000.000-00"         
    ClientsResponse:
      type: object
      properties:
        id:
          type: integer
        cpf:
          type: string
    ContactsRequest:
      type: object
      required:
        - client_id
        - type_contact
        - value_contact
      properties:
        client_id:
          type: integer
          example: 1
        type_contact:
          type: string
          enum:
            - email
            - telefone
          example: "email"
        value_contact:
          type: string
          maxLength: 100
          example: "tiago@email.com"     
    ContactsResponse:
      type: object
      properties:
        id:
          type: integer
        client_id:
          type: integer
        type_contact:
          type: string
        value_contact:
          type: string      
    PetsRequest:
      type: object
      required:
        - client_id
        - breed_id
        - birth_date
        - name
      properties:
        client_id:
          type: integer
          example: 1
        breed_id:
          type: integer
          example: 5
        birth_date:
          type: string
          format: date
          example: "2020-01-25"
        name:
          type: string
          maxLength: 50
          example: "Tommy"       
    PetsResponse:
      type: object
      properties:
        id:
          type: integer
        client_id:
          type: integer
        breed_id:
          type: integer
        birth_date:
          type: string
          format: data
        name:
          type: string
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer  